Create a ln(x+1) solving task
this is a program used for calculating the function of y=ln(x+1) and it contains two kinds of solutions. to make use of the other one you will just have to remove the /*and*/between line21 and line28. by the way make sure you have put line 29 and line 39 under "//"